+++ src/subjects/platform.rs
@@ -66,22 +66,16 @@ impl Subject for PlatformSubject {
 
 lazy_static! {
   static ref PLATFORM_EXPORT_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(EXPORT_COMMAND, None, "Export default platform configuration")
-      .set_long_about(
-        "Export the default platform configuration json file from the dsh-api library. \
+    CapabilityBuilder::new(EXPORT_COMMAND, None, &PlatformExport {}, "Export default platform configuration").set_long_about(
+      "Export the default platform configuration json file from the dsh-api library. \
         This file can be used as a starting point when platform customization is required."
-      )
-      .set_default_command_executor(&PlatformExport {})
+    )
   );
-  static ref PLATFORM_LIST_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), "List platforms")
-      .set_long_about("Lists all dsh platforms.")
-      .set_default_command_executor(&PLatformList {}),
-  );
+  static ref PLATFORM_LIST_CAPABILITY: Box<(dyn Capability + Send + Sync)> =
+    Box::new(CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), &PLatformList {}, "List platforms").set_long_about("Lists all dsh platforms."));
   static ref PLATFORM_OPEN_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(OPEN_COMMAND, Some(OPEN_COMMAND_ALIAS), "Open console or web application")
+    CapabilityBuilder::new(OPEN_COMMAND, Some(OPEN_COMMAND_ALIAS), &PlatformOpen {}, "Open console or web application")
       .set_long_about("Open the DSH console, monitoring page or the web application for the tenant or a service.")
-      .set_default_command_executor(&PlatformOpen {})
       .add_subcommands(vec![
         Command::new(OPEN_APP)
           .about("Open the console for the target platform/tenant and the provided app")
@@ -101,9 +95,8 @@ lazy_static! {
       ])
   );
   static ref PLATFORM_SHOW_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(SHOW_COMMAND, Some(SHOW_COMMAND_ALIAS), "Show platform data")
+    CapabilityBuilder::new(SHOW_COMMAND, Some(SHOW_COMMAND_ALIAS), &PlatformShow {}, "Show platform data")
       .set_long_about("Show platform data.")
-      .set_default_command_executor(&PlatformShow {})
       .add_target_argument(platform_name_argument())
       .add_extra_arguments(vec![
         app_id_argument().long("app"),