+++ src/subjects/target.rs
@@ -62,33 +62,28 @@ impl Subject for TargetSubject {
 
 lazy_static! {
   static ref TARGET_CREATE_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(CREATE_COMMAND, Some(CREATE_COMMAND_ALIAS), "Create a new target configuration")
+    CapabilityBuilder::new(CREATE_COMMAND, Some(CREATE_COMMAND_ALIAS), &TargetCreate {}, "Create a new target configuration")
       .set_long_about(
         "Create a new target configuration. \
         You will be prompted for the target's platform, tenant and password. \
         The platform and tenant will be stored in an unencrypted configuration file. \
         The password will be stored in your computer's keyring, which is more secure.",
       )
-      .set_default_command_executor(&TargetCreate {})
       .add_target_argument(platform_name_argument().required(true))
       .add_target_argument(tenant_name_argument().required(true))
   );
   static ref TARGET_DELETE_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(DELETE_COMMAND, None, "Delete target configuration")
+    CapabilityBuilder::new(DELETE_COMMAND, None, &TargetDelete {}, "Delete target configuration")
       .set_long_about(
         "Delete a target configuration. \
         You will be prompted for the target's platform and tenant, \
         and you need to explicitly confirm the action.",
       )
-      .set_default_command_executor(&TargetDelete {})
       .add_target_argument(platform_name_argument().required(true))
       .add_target_argument(tenant_name_argument().required(true))
   );
-  static ref TARGET_LIST_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), "List all target configurations")
-      .set_long_about("Lists all target configurations.")
-      .set_default_command_executor(&TargetList {})
-  );
+  static ref TARGET_LIST_CAPABILITY: Box<(dyn Capability + Send + Sync)> =
+    Box::new(CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), &TargetList {}, "List all target configurations").set_long_about("Lists all target configurations."));
   static ref TARGET_CAPABILITIES: Vec<&'static (dyn Capability + Send + Sync)> =
     vec![TARGET_CREATE_CAPABILITY.as_ref(), TARGET_DELETE_CAPABILITY.as_ref(), TARGET_LIST_CAPABILITY.as_ref()];
 }