+++ src/subjects/volume.rs
@@ -59,37 +59,32 @@ impl Subject for VolumeSubject {
 
 lazy_static! {
   static ref VOLUME_CREATE_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(CREATE_COMMAND, Some(CREATE_COMMAND_ALIAS), "Create new volume")
+    CapabilityBuilder::new(CREATE_COMMAND, Some(CREATE_COMMAND_ALIAS), &VolumeCreate {}, "Create new volume")
       .set_long_about("Create a new volume.")
-      .set_default_command_executor(&VolumeCreate {})
       .add_target_argument(volume_id_argument().required(true))
       .add_extra_argument(size_flag())
   );
   static ref VOLUME_DELETE_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(DELETE_COMMAND, None, "Delete volume")
+    CapabilityBuilder::new(DELETE_COMMAND, None, &VolumeDelete {}, "Delete volume")
       .set_long_about("Delete a volume.")
-      .set_default_command_executor(&VolumeDelete {})
       .add_target_argument(volume_id_argument().required(true))
   );
   static ref VOLUME_LIST_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), "List volumes")
+    CapabilityBuilder::new(LIST_COMMAND, Some(LIST_COMMAND_ALIAS), &VolumeListAll {}, "List volumes")
       .set_long_about("Lists all available volumes.")
-      .set_default_command_executor(&VolumeListAll {})
       .add_command_executors(vec![
         (FlagType::AllocationStatus, &VolumeListAllocationStatus {}, None),
         (FlagType::Configuration, &VolumeListConfiguration {}, None),
         (FlagType::Ids, &VolumeListIds {}, None),
         (FlagType::Usage, &VolumeListUsage {}, None),
       ])
-      .set_run_all_executors(true)
       .add_filter_flags(vec![
         (FilterFlagType::App, Some("List all apps that use the volume.".to_string())),
         (FilterFlagType::Service, Some("List all services that use the volume.".to_string()))
       ])
   );
   static ref VOLUME_SHOW_CAPABILITY: Box<(dyn Capability + Send + Sync)> = Box::new(
-    CapabilityBuilder::new(SHOW_COMMAND, Some(SHOW_COMMAND_ALIAS), "Show secret configuration")
-      .set_default_command_executor(&VolumeShowAll {})
+    CapabilityBuilder::new(SHOW_COMMAND, Some(SHOW_COMMAND_ALIAS), &VolumeShowAll {}, "Show secret configuration")
       .add_command_executors(vec![
         (FlagType::AllocationStatus, &VolumeShowAllocationStatus {}, None),
         (FlagType::Usage, &VolumeShowUsage {}, None),